Updated updated VectorCAST API for mocks (#169)
Move towards the updated API in VectorCAST for coded mocks
